added registration routes
